// src/app/order/interfaces/submit.ts
export interface Submit {
    countryValue: string;
    cityValue: string;
    regionValue: string;
    addressValue: string;
    zipValue: string;
    dateValue: string;
    nameValue: string;
    surnameValue: string;
    emailValue: string;
    descriptionValue: string;
    radioDateValue: string;
    radioPaymentValue: string;
}